pyproject.toml
[project]
name = "paladin-ai"
version = "0.1.0"
description = "AI-Powered Monitoring & Incident Response Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["python-dotenv"]
